feat: Load DJ specific content from all entities for individual page
Feature loads data for a specific DJ from Programs, Shows, and DJ collections and makes this accessible from the DJ controller through the findOne method. Angular uses findOne to make 3 separate calls to each entities modular, and the modulars verify that the query is for a DJ, and responds with the data that match query params.